Refactored finding cafes
