+++ src/App.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import './App.css';
+import { findCafes } from './utilities/findCafes';
 
 function App() {
   const [reviews, setReviews] = useState([]);
@@ -15,35 +16,39 @@ function App() {
   useEffect(() => {
   }, []);
 
-  const findCafes = () => {
-    const radiusMeters = parseFloat(dist) * 1609.34; 
+  // const findCafes = () => {
+  //   const radiusMeters = parseFloat(dist) * 1609.34; 
 
-    const geocoder = new window.google.maps.Geocoder();
-    geocoder.geocode({ address: address }, (results, status) => {
-      if (status === "OK" && results[0]) {
-        const location = results[0].geometry.location;
+  //   const geocoder = new window.google.maps.Geocoder();
+  //   geocoder.geocode({ address: address }, (results, status) => {
+  //     if (status === "OK" && results[0]) {
+  //       const location = results[0].geometry.location;
 
-        const service = new window.google.maps.places.PlacesService(document.createElement("div"));
-        service.nearbySearch(
-          {
-            location: location,
-            radius: radiusMeters,
-            type: "cafe",
-          },
-          (results, status) => {
-            if (status === window.google.maps.places.PlacesServiceStatus.OK) {
-              setCafes(results.map((place) => place.name));
-            } else {
-              alert("No cafes found within the specified radius.");
-              setCafes([]); 
-            }
-          }
-        );
-      } else {
-        alert("Invalid zip code or location not found.");
-        setCafes([]);
-      }
-    });
+  //       const service = new window.google.maps.places.PlacesService(document.createElement("div"));
+  //       service.nearbySearch(
+  //         {
+  //           location: location,
+  //           radius: radiusMeters,
+  //           type: "cafe",
+  //         },
+  //         (results, status) => {
+  //           if (status === window.google.maps.places.PlacesServiceStatus.OK) {
+  //             setCafes(results.map((place) => place.name));
+  //           } else {
+  //             alert("No cafes found within the specified radius.");
+  //             setCafes([]); 
+  //           }
+  //         }
+  //       );
+  //     } else {
+  //       alert("Invalid zip code or location not found.");
+  //       setCafes([]);
+  //     }
+  //   });
+  // };
+
+  const findCafesWrapper = () => {
+    findCafes(dist, address, setCafes);
   };
 
   const handleReviewSubmit = () => {
@@ -157,7 +162,7 @@ function App() {
             value={dist}
             onChange={(e) => setDist(e.target.value)}
           />
-          <button onClick={findCafes}>
+          <button onClick={findCafesWrapper}>
             Search
           </button>
